Use HotReloadDeltaGeneratorComputeScriptOutputs task in BuildTool.targets
